+++ src/Extensions.cs
@@ -1484,7 +1484,9 @@ public static class Extensions
   /// <summary>Apply friction to a projectile and update the cached baseData.speed</summary>
   public static void ApplyFriction(this Projectile p, float friction)
   {
-    p.baseData.speed *= (float)Lazy.FastPow(friction, p.LocalDeltaTime * C.FPS);
+    //WARNING: can't use FastPow here as it isn't accurate enough (e.g., slows down RC Launcher projectiles at higher frame raters)
+    // p.baseData.speed *= (float)Lazy.FastPow(friction, p.LocalDeltaTime * C.FPS);
+    p.baseData.speed *= Mathf.Pow(friction, p.LocalDeltaTime * C.FPS);
     p.UpdateSpeed();
   }